historybox.c/.h: GtkCombo -> GtkComboBoxText with entry
HistoryBox *hb = g_object_new(HISTORY_BOX_TYPE, "has-entry", TRUE, NULL);

The first thing that g_object_new() does is to sort the properties that you pass to it
into construct and non-construct properties. The list of construct properties is in
the order that you passed them. Added to this list (at the end) are any construct
properties that you did not pass to g_object_new(), with their default values.

http://blogs.gnome.org/desrt/2012/02/26/a-gentle-introduction-to-gobject-construction/